NuevasModificaciones

+++ scripts/mostrarCarrito.js
@@ -38,20 +38,38 @@ function activarClickEnBotones2() {
   let xCarrito = document.querySelectorAll('button.eliminarProducto');
   xCarrito.forEach((botonX) => {
     botonX.addEventListener('click', () => {
-      // Obtener el índice del producto a eliminar según el índice del botón
-      let indiceProducto = Array.from(xCarrito).indexOf(botonX);
+      Swal.fire({
+        title: 'Cuidado!',
+        text: "Usted quiere quitar el producto del carrito?",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si borrar!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Eliminado',
+            'El producto se borro del carrito',
+            'success'
+          )
+          let indiceProducto = Array.from(xCarrito).indexOf(botonX);
         
-      if (indiceProducto !== -1) {
-        // Eliminar el producto del carrito usando el índice
-        carrito.splice(indiceProducto, 1);
-        localStorage.setItem('Carrito', JSON.stringify(carrito))
-          
-  
-        // Volver a armar el carrito para reflejar los cambios en la tabla
-        armarCarrito();
-        finalizarCompra()
-        activarClickEnBotones2()
-      }
+          if (indiceProducto !== -1) {
+            // Eliminar el producto del carrito usando el índice
+            carrito.splice(indiceProducto, 1);
+            localStorage.setItem('Carrito', JSON.stringify(carrito))
+              
+      
+            // Volver a armar el carrito para reflejar los cambios en la tabla
+            armarCarrito();
+            finalizarCompra()
+            activarClickEnBotones2()
+          } 
+        }
+      })
+      // Obtener el índice del producto a eliminar según el índice del botón
+      
     });
   });
 }
